@@ -14,36 +14,36 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "8.0.3",
-    "@angular/common": "8.0.3",
-    "@angular/compiler": "8.0.3",
-    "@angular/core": "8.0.3",
-    "@angular/forms": "8.0.3",
-    "@angular/material": "8.0.1",
-    "@angular/platform-browser": "8.0.3",
-    "@angular/platform-browser-dynamic": "8.0.3",
-    "@angular/router": "8.0.3",
+    "@angular/animations": "8.1.1",
+    "@angular/common": "8.1.1",
+    "@angular/compiler": "8.1.1",
+    "@angular/core": "8.1.1",
+    "@angular/forms": "8.1.1",
+    "@angular/material": "8.0.2",
+    "@angular/platform-browser": "8.1.1",
+    "@angular/platform-browser-dynamic": "8.1.1",
+    "@angular/router": "8.1.1",
     "rxjs": "6.5.2",
     "tslib": "1.10.0",
     "zone.js": "0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.800.6",
-    "@angular-devkit/build-ng-packagr": "0.800.6",
-    "@angular/cdk": "8.0.1",
-    "@angular/cli": "8.0.6",
-    "@angular/compiler-cli": "8.0.3",
-    "@angular/language-service": "8.0.3",
+    "@angular-devkit/build-angular": "0.801.1",
+    "@angular-devkit/build-ng-packagr": "0.801.1",
+    "@angular/cdk": "8.0.2",
+    "@angular/cli": "8.1.1",
+    "@angular/compiler-cli": "8.1.1",
+    "@angular/language-service": "8.1.1",
     "@types/jasmine": "3.3.13",
     "@types/jasminewd2": "2.0.6",
-    "@types/node": "12.0.10",
+    "@types/node": "12.6.2",
     "codelyzer": "5.1.0",
     "core-js": "3.1.4",
     "cpx": "1.5.0",
     "jasmine-core": "3.4.0",
     "jasmine-spec-reporter": "4.2.1",
-    "karma": "4.1.0",
-    "karma-chrome-launcher": "2.2.0",
+    "karma": "4.2.0",
+    "karma-chrome-launcher": "3.0.0",
     "karma-coverage-istanbul-reporter": "2.0.5",
     "karma-jasmine": "2.0.1",
     "karma-jasmine-html-reporter": "1.4.2",
